# server/swagger_output.json
{
  "openapi": "3.1.0",
  "info": {
    "version": "v1.0.0",
    "title": "Co-Work Project",
    "description": "Implementation of Swagger with TypeScript"
  },
  "servers": [
    {
      "url": "https://handsomelai.shop",
      "description": ""
    }
  ],
  "paths": {
...
    "/api/user/signup": {
      "post": {
        "tags": ["User"],
        "summary": "sign up the user account",
        "description": "record the user information in the database, and return JWT token for autthentication",
        "responses": {
          "200": {
            "description": "sign up successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "data": {
                      "type": "object",
                      "properties": {
                        "access_token": {
                          "type": "string",
                          "example": "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJ1c2VySWQiOjI1LCJpYXQiOjE2OTkwODc2MzEsImV4cCI6MTY5OTA5MTIzMX0.Mh16vZAmi_5badOHesHY4RZHkIc3jRmYRvLv5aDC72I"
                        },
                        "access_expired": {
                          "type": "integer",
                          "example": 3600
                        },
                        "user": {
                          "$ref": "#/components/schemas/User"
                        }
                      }
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Invalid input",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "errors": {
                      "type": "array",
                      "items": {
                        "$ref": "#/components/schemas/BadRequest-invalid-value"
                      }
                    }
                  }
                }
              }
            }
          },
          "500": {
            "description": "Internal Server Error"
          }
        },
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "name": {
                    "example": "Ian"
                  },
                  "email": {
                    "example": "Ian@gmail.com"
                  },
                  "password": {
                    "example": "1234"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api/user/signin": {
      "post": {
        "tags": ["User"],
        "summary": "sign in by the user account",
        "description": "check the user information in the database, and return JWT token for autthentication",
        "responses": {
          "200": {
            "description": "sign up successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "data": {
                      "type": "object",
                      "properties": {
                        "access_token": {
                          "type": "string",
                          "example": "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJ1c2VySWQiOjI1LCJpYXQiOjE2OTkwODc2MzEsImV4cCI6MTY5OTA5MTIzMX0.Mh16vZAmi_5badOHesHY4RZHkIc3jRmYRvLv5aDC72I"
                        },
                        "access_expired": {
                          "type": "integer",
                          "example": 3600
                        },
                        "user": {
                          "$ref": "#/components/schemas/User"
                        }
                      }
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Invalid input",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "errors": {
                      "type": "array",
                      "items": {
                        "$ref": "#/components/schemas/BadRequest-invalid-value"
                      }
                    }
                  }
                }
              }
            }
          }
        },
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "provider": {
                    "example": "native"
                  },
                  "name": {
                    "example": "Ian"
                  },
                  "email": {
                    "example": "Ian@gmail.com"
                  },
                  "password": {
                    "example": "1234"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api/user/profile": {
      "get": {
        "tags": ["User"],
        "summary": "get user information",
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "successful operation",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "data": {
                      "$ref": "#/components/schemas/User"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Bad Request"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "User": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "example": "25"
          },
          "provider": {
            "type": "string",
            "example": "native"
          },
          "name": {
            "type": "string",
            "example": "Ian"
          },
          "email": {
            "type": "string",
            "example": "ian@gmail.com"
          },
          "picture": {
            "type": "string",
            "example": ""
          }
        }
      },
...
      "BadRequest-invalid-value": {
        "type": "string",
        "example": "query[paging]: Invalid value"
      },
...
    },
    "securitySchemes": {
      "BearerAuth": {
        "description": "Please enter JWT token",
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    }
  },
...
}